% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demark.R
\name{demark}
\alias{demark}
\title{Function for use in PMark function}
\usage{
demark(formula, data, n, cut_p, prior, replace = F)
}
\arguments{
\item{formula}{formula.}

\item{data}{data frame containing the variables used in the formula.}

\item{n}{the number of cases to subsample from for random sampling}

\item{cut_p}{Numeric. A number between 0 and 1 indicating the probability for the new cutoff points.}

\item{prior}{A vector containing the desired prior probability for each group. Default is an uninformative prior.}
}
\description{
This function is used by PMark to calculate the cutoff points based on a given probability
}
