% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBhum.R
\docType{data}
\name{MBhum}
\alias{MBhum}
\title{Whole dataset for Middenbeemster.}
\format{
Data frame with 106 rows and 7 variables.
\describe{
 \item{[,1] ID}{Individual identifier}

 \item{[,2] Sex}{Biological sex of each individual}

 \item{[,3] MaxL}{Humeral max length measurement}

 \item{[,4] HeadD}{Humeral head diameter measurement}

 \item{[,5] EpiB}{Humeral epicondylar breadth measurement}

}
}
\usage{
MBhum
}
\description{
The data for all 106 individuals from Middenbeemster used in the analysis.
}
\keyword{datasets}
