% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMark.R, R/PMark.formula.R
\name{pmark}
\alias{pmark}
\alias{PMark.formula}
\title{Probability demarking points for linear discriminant analysis}
\usage{
pmark(object, ...)

PMark.formula(
  formula,
  data,
  cut_p,
  iter = 500,
  prior = c(0.5, 0.5),
  replace = T,
  n = NULL
)
}
\arguments{
\item{object}{Object of class lda}

\item{...}{arguments passed to methods.}

\item{formula}{formula. groups ~ x1 + x2 + ... + xn.}

\item{data}{data frame from which the variables from the formula will be extracted.}

\item{cut_p}{numeric. Probability of desired PMarks between 0 and 1. Recommended: 0.8, 0.95, or 0.99.}

\item{iter}{number of iterations for bootstrapping. Recommended: >= 500. If the number of iterations exceeds the number of different combinations of subsamples, the function will issue a warning.}

\item{prior}{A vector containing the desired prior probability for each group. Default is an uninformative prior.}

\item{replace}{Whether to use  replacement (sub-sampling) or not (bootstrap entire sample).}

\item{n}{size of the subsample for random sampling. Must not exceed the total sample size.}
}
\value{
An object of class "pmark", which is identical to an "lda" object, but with $PMark added to the list,
which contains the computed PMark values for each group.
}
\description{
Probability demarking points for linear discriminant analysis
}
\details{
This function calculates cutoff points for linear discriminant analysis (lda) using posterior probabilities. The output will be a discriminant score associated with the specified posterior probability found by bootstrapping.
}
\examples{
#PMark example
#example using a formula (and 95\% PMarks):
mark1 <- pmark(Sex ~ HeadD, MBhum, 0.95)
mark1$PMark
#example using an lda object and resampling without replacement.
dfa1 <- MASS::lda(Sex ~ HeadD + EpiB, MBhum)
mark2 <- pmark(dfa1, cut_p = 0.8, replace = FALSE, n = 80)
mark2$PMark
}
\seealso{
\code{\link[MASS]{lda}}, \code{\link[MASS]{predict.lda}}
}
