% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMark_shiny.R
\name{PMark_shiny}
\alias{PMark_shiny}
\title{Function used to compute PMarks in the shiny app}
\usage{
PMark_shiny(formula, data = "MB11", n, cut_p, iter, prior)
}
\arguments{
\item{formula}{formula to be used to create the linear discrminant model}

\item{data}{Specifies if the data should come from the built-in MB11 dataset or if it should be an external dataset. If not "MB11", upload a data frame.}

\item{n}{size of the subsample for random sampling. Must be smaller than the total sample size.}

\item{cut_p}{numeric. Probability of desired PMarks between 0 and 1. Recommended: 0.8, 0.95, or 0.99.}

\item{iter}{number of iterations for bootstrapping. Recommended: >= 500. If the number of iterations exceeds the number of different combinations of subsamples, the function will issue a warning.}

\item{prior}{A vector containing the desired prior probability for each group. Default is an uninformative prior.}
}
\value{
Returns the PMark value and a plot of the analysis including the PMark value.
}
\description{
Function used to compute PMarks in the shiny app
}
