% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_sex.R
\name{pred_sex}
\alias{pred_sex}
\title{Function to estimate sex in the shiny app}
\usage{
pred_sex(MaxL = NA, HeadD = NA, EpiB = NA, prior)
}
\arguments{
\item{MaxL}{Max humeral length measurement (mm)}

\item{HeadD}{Humeral head diameter measurement (mm)}

\item{EpiB}{Humeral epicondylar breadth measurement (mm)}

\item{prior}{Prior probability to be passed on to the lda and qda functions}
}
\value{
Returns a data frame with the results of the analyses. The columns represent the classification and probability. The rows represent the logit, probit, lda, and qda methods, respectively.
}
\description{
Sex estimation based on the Middenbeemster reference sample using logit, probit, lda, and qda.
}
