% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.pmark.R
\name{print.pmark}
\alias{print.pmark}
\title{Printing method for pmark funcitons.}
\usage{
\method{print}{pmark}(x)
}
\arguments{
\item{x}{object of type 'pmark'.}
}
\description{
Printing method for pmark funcitons.
}
