% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPMark.R
\name{runPMark}
\alias{runPMark}
\title{Function to launch the pmarkr shinyapp}
\usage{
runPMark()
}
\description{
This function launches the shiny graphical user interface (GUI) for the pmarkr package.
}
\details{
The GUI is provided as a user-friendly alternative for less experienced R-users. It can be used to obtain the PMarks from the MB11 sample or an uploaded dataset.
}
\examples{
##Not run: runPMark()
}
