% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMark.R, R/PMark_default.R, R/PMark_formula.R,
%   R/PMark_lda.R
\name{PMark}
\alias{PMark}
\alias{PMark.default}
\alias{PMark.formula}
\alias{PMark.lda}
\title{Probability demarking points for linear discriminant analysis}
\usage{
PMark(x, data, n, cut_p, iter, prior = c(0.5, 0.5))

\method{PMark}{default}(formula, data, n, cut_p, iter, prior = c(0.5,
  0.5))

\method{PMark}{formula}(formula, data, n, cut_p, iter, prior = c(0.5,
  0.5))

\method{PMark}{lda}(object, data = NULL, n, cut_p, iter, prior = c(0.5,
  0.5))
}
\arguments{
\item{x}{either a formula or an lda object}

\item{data}{data frame from which the variables from the formula will be extracted.}

\item{n}{size of the subsample for random sampling. Must be smaller than the total sample size.}

\item{cut_p}{numeric. Probability of desired PMarks between 0 and 1. Recommended: 0.8, 0.95, or 0.99.}

\item{iter}{number of iterations for bootstrapping. Recommended: >= 500. If the number of iterations exceeds the number of different combinations of subsamples, the function will issue a warning.}

\item{prior}{A vector containing the desired prior probability for each group. Default is an uninformative prior.}

\item{formula}{formula. groups ~ x1 + x2 + ... + xn.}

\item{object}{Object of class lda}
}
\value{
An object of class "pmark", which is identical to an "lda" object, but with $PMark added to the list, which contains the computed PMark value.
}
\description{
This function calculates cutoff points for linear discriminant analysis (lda) using posterior probabilities. The output will be a discriminant score associated with the specified posterior probability found by bootstrapping.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Probability demarking points for linear discriminant analysis using a formula

\item \code{formula}: Probability demarking points for linear discriminant analysis using a formula

\item \code{lda}: Probability demarking points for linear discriminant analysis using an object of class 'lda'.
}}

\examples{
#PMark example
#example using a formula:
mark1 <- PMark(Sex ~ HeadD, MBhum, 80, 0.95, 500, prior = c(0.5,0.5))
mark1$PMark
#example using an lda object
dfa1 <- MASS::lda(Sex ~ HeadD + EpiB, MBhum)
mark2 <- PMark(dfa1, n = 80, cut_p = 0.8, iter = 500)
mark2$PMark
}
\seealso{
\code{\link[MASS]{lda}}, \code{\link[MASS]{predict.lda}}
}
